# internal/api/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lybrix control plane",
    "description": "Document Ingestion & Retrieval Platform \u2014 PRD docs/prd.md",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": ["models"],
        "summary": "List Models",
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["models"],
        "summary": "Create Model",
        "operationId": "create_model_v1_models_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOut"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate name",
            "content": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/models/{model_id}": {
      "post": {
        "tags": ["models"],
        "summary": "Update Model",
        "operationId": "update_model_v1_models__model_id__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOut"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "409": {
            "description": "Name clash",
            "content": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["models"],
        "summary": "Delete Model",
        "operationId": "delete_model_v1_models__model_id__delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "409": {
            "description": "Bound or default",
            "content": {}
          }
        }
      }
    },
    "/v1/collections/{collection_id}/model": {
      "post": {
        "tags": ["models"],
        "summary": "Bind Collection Model",
        "operationId": "bind_collection_model_v1_collections__collection_id__model_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "ModelOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "enum": ["tei", "ollama", "openai"],
            "title": "Provider"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "ingest_url": {
            "type": "string",
            "title": "Ingest Url"
          },
          "query_url": {
            "type": "string",
            "title": "Query Url"
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key"
          },
          "vector_dim": {
            "type": "integer",
            "maximum": 2000,
            "minimum": 1,
            "title": "Vector Dim"
          },
          "query_prefix": {
            "type": "string",
            "title": "Query Prefix",
            "default": "search_query: "
          },
          "batch_size": {
            "type": "integer",
            "minimum": 1,
            "title": "Batch Size",
            "default": 48
          },
          "ctx_budget": {
            "type": "integer",
            "minimum": 1,
            "title": "Ctx Budget",
            "default": 1900
          },
          "truncate_chars": {
            "type": "integer",
            "minimum": 0,
            "title": "Truncate Chars",
            "default": 6000
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": ["id", "name", "provider", "model_id", "ingest_url", "query_url", "vector_dim"],
        "title": "ModelOut"
      },
      "ModelCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "enum": ["tei", "ollama", "openai"],
            "title": "Provider"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "ingest_url": {
            "type": "string",
            "title": "Ingest Url"
          },
          "query_url": {
            "type": "string",
            "title": "Query Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "vector_dim": {
            "type": "integer",
            "maximum": 2000,
            "minimum": 1,
            "title": "Vector Dim"
          },
          "query_prefix": {
            "type": "string",
            "title": "Query Prefix",
            "default": "search_query: "
          },
          "batch_size": {
            "type": "integer",
            "minimum": 1,
            "title": "Batch Size",
            "default": 48
          },
          "ctx_budget": {
            "type": "integer",
            "minimum": 1,
            "title": "Ctx Budget",
            "default": 1900
          },
          "truncate_chars": {
            "type": "integer",
            "minimum": 0,
            "title": "Truncate Chars",
            "default": 6000
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": ["name", "provider", "model_id", "ingest_url", "query_url", "vector_dim"],
        "title": "ModelCreate"
      },
      "ModelUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1,
                "title": "Name"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["tei", "ollama", "openai"],
                "title": "Provider"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string",
                "title": "Model Id"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "ingest_url": {
            "anyOf": [
              {
                "type": "string",
                "title": "Ingest Url"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Url"
          },
          "query_url": {
            "anyOf": [
              {
                "type": "string",
                "title": "Query Url"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "null = unchanged"
          },
          "vector_dim": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000,
                "minimum": 1,
                "title": "Vector Dim"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Dim"
          },
          "query_prefix": {
            "anyOf": [
              {
                "type": "string",
                "title": "Query Prefix",
                "default": "search_query: "
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Prefix"
          },
          "batch_size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "title": "Batch Size",
                "default": 48
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size"
          },
          "ctx_budget": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "title": "Ctx Budget",
                "default": 1900
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctx Budget"
          },
          "truncate_chars": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "title": "Truncate Chars",
                "default": 6000
              },
              {
                "type": "null"
              }
            ],
            "title": "Truncate Chars"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean",
                "title": "Is Default",
                "default": false
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "type": "object",
        "title": "ModelUpdate"
      },
...
    }
  }
}